FISH-12188 Concurro 3.1.0

+++ appserver/concurrent/concurrent-impl/src/main/java/org/glassfish/concurrent/runtime/deployer/ManagedExecutorServiceConfig.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-// Portions Copyright [2022-2024] Payara Foundation and/or affiliates
+// Portions Copyright 2022-2025 Payara Foundation and/or its affiliates
 package org.glassfish.concurrent.runtime.deployer;
 
 import org.glassfish.concurrent.config.ManagedExecutorService;
@@ -56,7 +56,7 @@ public class ManagedExecutorServiceConfig extends BaseConfig  {
     private long keepAliveSeconds;
     private int maximumPoolSize;
     private int taskQueueCapacity;
-    private long threadLifeTimeSeconds;
+    private int threadLifeTimeSeconds;
     private String context;
 
     public ManagedExecutorServiceConfig(ManagedExecutorService config) {
@@ -70,7 +70,7 @@ public class ManagedExecutorServiceConfig extends BaseConfig  {
         keepAliveSeconds = parseLong(config.getKeepAliveSeconds(), 60);
         maximumPoolSize = parseInt(config.getMaximumPoolSize(), Integer.MAX_VALUE);
         taskQueueCapacity = parseInt(config.getTaskQueueCapacity(), Integer.MAX_VALUE);
-        threadLifeTimeSeconds = parseLong(config.getThreadLifetimeSeconds(), 0L);
+        threadLifeTimeSeconds = parseInt(config.getThreadLifetimeSeconds(), 0);
         context = config.getContext();
     }
 
@@ -102,7 +102,7 @@ public class ManagedExecutorServiceConfig extends BaseConfig  {
         return taskQueueCapacity;
     }
 
-    public long getThreadLifeTimeSeconds() {
+    public int getThreadLifeTimeSeconds() {
         return threadLifeTimeSeconds;
     }
